Added exit on failure to find .smali launcher
Instead of running through the whole code and being told you are good, I had it exit upon failure to grep the .smali file to be used for pointing the injected payload. Currently, the script has worked for 2 of 3 apps. The grep line could be modified as needed. I may add a user input entry for when it fails to auto-find allowing the user to type in the .smali path they manually locate in the AndroidManifest.xml file.