Added a few checks for failures
Added checks to exit the script with errors if needed instead of telling the user it is finished when it didn't even compile. Checks include verifying the .apk file is present after apktool b is ran. Also, when manually entering the path it checks to see if the .smali file is present before attempting to inject into it.