Infinite loop bug fix
The until loop was set to exit on "-eq 1", but if unable to find the line number  on line 62 this caused an infinite loop. Originally, it was believes the LAUNCHER was located directly after MAIN in the AndroidManifest.xml file every time, but this was found not to always be the case. If this happens, the user will need to simply look at the grep statement and put in the path, such as: smali/com/path1/path2/Main.smali   .. verified the bug is fixed