Checks for insertion of invoke command
Added a check to verify the insertion of the invoke for the payload is in place, otherwise it will exit with an explanation of how to manually finish the project. The reason for this is because not every app uses the standard onCreate method, but this will halt cleanup and allow the user to manually place the invoke method and finishing compiling+signing. Also, added more print statements and spacing to make the output easier. Further, removed the verbose command from the jarsigner to not flood the screen as much to allow the user to easily debug.